Add advertiser sample
The advertiser sample configures a peripheral device to
advertise a custom payload and a central device to
scan for the peripheral device.
The custom advertisement payload contains a 1 byte
counter value that is incremented every two seconds.
The central device prints the counter value when it
receives a new advertisement.

L220D100-7
L200D100-17

Signed-off-by: Ryan Erickson <ryan.erickson@lairdconnect.com>