Add custom GATT sample
The custom GATT sample configures a peripheral
device to advertise a custom
payload
and sets up the BT SIG GATT Battery Service. Every 5 seconds, the peripheral will
change
the battery level.
A central device scans for the peripheral device
and connects to it.
The central device will received the battery level
every time it is updated.

L220D100-52